requirements.txt
python-telegram-bot==20.8
aiogram==2.25.1
python-dotenv==1.0.0
schedule
python-telegram-bot[job-queue]
python-jenkins